move code into a lib for better integration testing and start writing cucumber tests.
